Further changes to build system Now includes an entry for building the docs with latex
