+++ doc/NISE3.1.tex
