+++ doc/tableofcontents.tex
