+++ doc/techniques.tex
