--- findFFTW.cmake
@@ -1,16 +1,0 @@
-cmake_minimum_required(VERSION 2.8.2)
-
-project(findFFTW NONE)
-
-include(ExternalProject)
-
-ExternalProject_Add(findFFTW
-    GIT_REPOSITORY  "https://github.com/egpbos/findfftw.git"
-    CONFIGURE_COMMAND ""
-    BUILD_COMMAND ""
-    INSTALL_COMMAND ""
-    TEST_COMMAND ""
-    SOURCE_DIR "${CMAKE_CURRENT_BINARY_DIR}/findFFTW-src"
-    BINARY_DIR ""
-    INSTALL_DIR ""
-)
