+++ src/absorption.h
