+++ src/c_absorption.c
@@ -5,17 +5,16 @@
 #include <time.h>
 #include <fftw3.h>
 #include "omp.h"
-#include "types3.1.h"
-#include "NISE3.1subs.h"
-#include "absorption.h"
-#include "luminescence.h"
+#include "types.h"
+#include "NISE_subs.h"
+#include "c_absorption.h"
 #include "1DFFT.h"
 
-void luminescence(t_non *non){
+void c_absorption(t_non *non){
   // Initialize variables
   float *re_S_1,*im_S_1; // The first-order response function
   float *mu_eg,*Hamil_i_e;
-
+  float *mu_xyz;
   // Aid arrays
   float *vecr,*veci,*vecr_old,*veci_old;
 
@@ -27,7 +26,9 @@ void luminescence(t_non *non){
 
   /* File handles */
   FILE *H_traj,*mu_traj;
+  FILE *C_traj;
   FILE *outone,*log;
+  FILE *Cfile;
 
   /* Integers */
   int nn2;
@@ -36,17 +37,16 @@ void luminescence(t_non *non){
   int x,ti,tj,i;
   int t1,fft;
   int elements;
+  int cl,Ncl;
 
   /* Time parameters */
   time_t time_now,time_old,time_0;
-  fft=0; 
   /* Initialize time */
   time(&time_now);
   time(&time_0);
   shift1=(non->max1+non->min1)/2;
   printf("Frequency shift %f.\n",shift1);
   non->shifte=shift1;
-  printf("Temperature %f.\n",non->temperature);
 
   // Allocate memory
   re_S_1=(float *)calloc(non->tmax,sizeof(float));
@@ -66,8 +66,21 @@ void luminescence(t_non *non){
     printf("Dipole file %s not found!\n",non->dipoleFName);
     exit(1);
   }
+  //printf("A\n");
+  /* Open file with cluster information if appicable */
+  if (non->cluster!=-1){
+    Cfile=fopen("Cluster.bin","rb");
+    if (Cfile==NULL){
+      printf("Cluster option was activated but no Cluster.bin file provided.\n");
+      printf("Please, provide cluster file or remove Cluster keyword from\n");
+      printf("input file.\n");
+      exit(0);
+    }
+    Ncl=0; // Counter for snapshots calculated
+  }
 
   // Here we want to call the routine for checking the trajectory files
+  control(non);
 
   itime=0;
   // Do calculation
@@ -92,18 +105,58 @@ void luminescence(t_non *non){
   vecr_old=(float *)calloc(non->singles,sizeof(float));
   veci_old=(float *)calloc(non->singles,sizeof(float));
   mu_eg=(float *)calloc(non->singles,sizeof(float));
+  mu_xyz=(float *)calloc(non->singles*3,sizeof(float));
 
-  // Loop over samples
+  /* Read coupling */
+  if (!strcmp(non->hamiltonian,"Coupling")){
+    C_traj=fopen(non->couplingFName,"rb");
+    if (C_traj==NULL){
+      printf("Coupling file not found!\n");
+      exit(1);
+    }
+    if (read_He(non,Hamil_i_e,C_traj,-1)!=1){
+      printf("Coupling trajectory file to short, could not fill buffer!!!\n");
+      exit(1);
+    }
+    fclose(C_traj);
+    for (x=0;x<3;x++){
+      if (read_mue(non,mu_xyz+non->singles*x,mu_traj,0,x)!=1){
+         printf("Dipole trajectory file to short, could not fill buffer!!!\n");
+         printf("ITIME %d %d\n",0,x);
+         exit(1);
+      }
+    }
+  }
+
+  /* Loop over samples */
   for (samples=non->begin;samples<non->end;samples++){
 
-    // Calculate linear response    
+    /* Calculate linear response */   
     ti=samples*non->sample;
-    for (x=0;x<3;x++){
-      // Read mu(ti)
-      if (read_mue(non,vecr,mu_traj,ti,x)!=1){
-	printf("Dipole trajectory file to short, could not fill buffer!!!\n");
-	printf("ITIME %d %d\n",ti,x);
+    if (non->cluster!=-1){
+      if (read_cluster(non,ti,&cl,Cfile)!=1){
+	printf("Cluster trajectory file to short, could not fill buffer!!!\n");
+	printf("ITIME %d\n",ti);
 	exit(1);
+      }
+      //      printf("%d\n",cl);
+      // Configuration belong to cluster
+      if (non->cluster==cl){
+	Ncl++;
+      }
+    }
+    if (non->cluster==-1 || non->cluster==cl){
+      
+    for (x=0;x<3;x++){
+      /* Read mu(ti) */
+      if (!strcmp(non->hamiltonian,"Coupling")){
+        copyvec(mu_xyz+non->singles*x,vecr,non->singles);
+      } else {
+        if (read_mue(non,vecr,mu_traj,ti,x)!=1){
+	  printf("Dipole trajectory file to short, could not fill buffer!!!\n");
+	  printf("ITIME %d %d\n",ti,x);
+	  exit(1);
+        }
       }
       clearvec(veci,non->singles);
       copyvec(vecr,vecr_old,non->singles);
@@ -111,29 +164,37 @@ void luminescence(t_non *non){
       // Loop over delay
       for (t1=0;t1<non->tmax;t1++){
 	tj=ti+t1;
-	// Read Hamiltonian
-	if (read_He(non,Hamil_i_e,H_traj,tj)!=1){
-	  printf("Hamiltonian trajectory file to short, could not fill buffer!!!\n");
-	  exit(1);
-	}
+	/* Read Hamiltonian */
+	if (!strcmp(non->hamiltonian,"Coupling")){
+          if (read_Dia(non,Hamil_i_e,H_traj,tj)!=1){
+            printf("Hamiltonian trajectory file to short, could not fill buffer!!!\n");
+            exit(1);
+          }
+        } else {
+	  if (read_He(non,Hamil_i_e,H_traj,tj)!=1){
+	    printf("Hamiltonian trajectory file to short, could not fill buffer!!!\n");
+	    exit(1);
+  	  }
+        }
 	
-	// Read mu(tj)
-	if (read_mue(non,mu_eg,mu_traj,tj,x)!=1){
-	  printf("Dipole trajectory file to short, could not fill buffer!!!\n");
-	  printf("JTIME %d %d\n",tj,x);
-	  exit(1);
-	}
+	/* Read mu(tj) */
+        if (!strcmp(non->hamiltonian,"Coupling")){
+          copyvec(mu_xyz+non->singles*x,mu_eg,non->singles);
+        } else {
+	  if (read_mue(non,mu_eg,mu_traj,tj,x)!=1){
+	    printf("Dipole trajectory file to short, could not fill buffer!!!\n");
+	    printf("JTIME %d %d\n",tj,x);
+	    exit(1);
+	  }
+        }
 
 	// Do projection on selected sites if asked
 	if (non->Npsites>0){
 	  projection(mu_eg,non);
 	}
-
-	// Add Boltzman weight
-	bltz_weight(mu_eg,Hamil_i_e,non);
 	
 	// Find response
-	calc_S1(re_S_1,im_S_1,t1,non,vecr,veci,mu_eg);
+	c_calc_S1(re_S_1,im_S_1,t1,non,vecr,veci,mu_eg);
 
 	// Probagate vector
 	if (non->propagation==1) propagate_vec_coupling_S(non,Hamil_i_e,vecr,veci,non->ts,1);
@@ -155,10 +216,12 @@ void luminescence(t_non *non){
 	}
       }
     }
+    } // Cluster loop
   
     // Log time
     log=fopen("NISE.log","a");
     fprintf(log,"Finished sample %d\n",samples);
+          
     time_now=log_time(time_now,log);
     fclose(log);
   }
@@ -168,6 +231,7 @@ void luminescence(t_non *non){
   free(vecr_old);
   free(veci_old);
   free(mu_eg);
+  free(mu_xyz);
   free(Hamil_i_e);
 
   // The calculation is finished, lets write output
@@ -176,99 +240,44 @@ void luminescence(t_non *non){
   fprintf(log,"Writing to file!\n");  
   fclose(log);
 
-  fclose(mu_traj),fclose(H_traj);
+  if (non->cluster!=-1){
+    printf("Of %d samples %d belonged to cluster %d.\n",samples,Ncl,non->cluster);
+    //    printf("Changing normalization to number of samples used.");
+    //    samples=Ncl;
+    if (samples==0){ // Avoid dividing by zero
+      samples=1;
+    }
+  }
 
-  outone=fopen("RLum.dat","w");
+  fclose(mu_traj),fclose(H_traj);
+  if (non->cluster!=-1){
+    fclose(Cfile);
+  }
+
+  /* Save time domain response */
+  outone=fopen("TD_Absorption.dat","w");
   for (t1=0;t1<non->tmax1;t1+=non->dt1){
     fprintf(outone,"%f %e %e\n",t1*non->deltat,re_S_1[t1]/samples,im_S_1[t1]/samples);
   }
   fclose(outone);
 
   /* Do Forier transform and save */
-  do_1DFFT(non,"Luminescence.dat",re_S_1,im_S_1,samples);
+  do_1DFFT(non,"Absorption.dat",re_S_1,im_S_1,samples);
 
   free(re_S_1),free(im_S_1);
 
   printf("----------------------------------------------\n");
-  printf(" Luminescence calculation succesfully completed\n");
+  printf(" Absorption calculation succesfully completed\n");
   printf("----------------------------------------------\n\n");
 
   return;
 }	
 
-void calc_LUM(float *re_S_1,float *im_S_1,int t1,t_non *non,float *cr,float *ci,float *mu){
+void c_calc_S1(float *re_S_1,float *im_S_1,int t1,t_non *non,float *cr,float *ci,float *mu){
   int i;
   for (i=0;i<non->singles;i++){
     re_S_1[t1]+=mu[i]*cr[i];
     im_S_1[t1]+=mu[i]*ci[i];
   }
-  return;
-}
-
-void bltz_weight(float *mu_eg,float *Hamiltonian_i,t_non *non){
-  int index,N;
-  float *H,*e,*c2,*c1;
-  float *cnr;
-  float *crr;
-  N=non->singles;
-  H=(float *)calloc(N*N,sizeof(float));
-  e=(float *)calloc(N,sizeof(float));
-  c2=(float *)calloc(N,sizeof(float));
-  c1=(float *)calloc(N,sizeof(float));
-  cnr=(float *)calloc(N*N,sizeof(float));
-  crr=(float *)calloc(N*N,sizeof(float));
-  int a,b,c;
-  float kBT=non->temperature*0.695; // Kelvin to cm-1
-  float Q,iQ;
-
-  // Build Hamiltonian
-  for (a=0;a<N;a++){
-    H[a+N*a]=Hamiltonian_i[a+N*a-(a*(a+1))/2]; // Diagonal
-    for (b=a+1;b<N;b++){
-      H[a+N*b]=Hamiltonian_i[b+N*a-(a*(a+1))/2];
-      H[b+N*a]=Hamiltonian_i[b+N*a-(a*(a+1))/2];
-    }
-  }
-  diagonalizeLPD(H,e,N);
- 
-  // Exponentiate [U=exp(-H/kBT)]
-  for (a=0;a<N;a++){
-    c2[a]=exp(-e[a]/kBT);
-    Q=Q+c2[a];
-  }
-  iQ=1.0/Q;
-
-  // Transform to site basis
-  for (a=0;a<N;a++){
-    for (b=0;b<N;b++){
-      cnr[b+a*N]+=H[b+a*N]*c2[b]*iQ;
-    }
-  }  
-  for (a=0;a<N;a++){
-    for (b=0;b<N;b++){
-      for (c=0;c<N;c++){
-        crr[a+c*N]+=H[b+a*N]*cnr[b+c*N];
-      }
-    }
-  }
-  // The weights in the site basis were calculated
-
-  // Multiply weights to dipole operator
-  for (a=0;a<N;a++){
-    for (b=0;b<N;b++){
-      c1[a]+=crr[a+b*N]*mu_eg[b];
-    }
-  }
-
-  // Update dipole operator
-  for (a=0;a<N;a++){
-    mu_eg[a]=c1[a];
-  }
-  free(H);
-  free(c2);
-  free(c1);
-  free(e);
-  free(crr);
-  free(cnr);
   return;
 }
