+++ src/c_absorption.h
