+++ src/luminescence.h
