+++ src/translate.h
